Move keywordCount to the abstract researcher
