+++ packages/yoastseo/src/languageProcessing/researches/keywordCount.js
@@ -1,10 +1,10 @@
 /** @module analyses/getKeywordCount */
 
-import matchWords from "../helpers/matchTextWithArray";
+import matchWords from "../helpers/match/matchTextWithArray";
 import { uniq as unique } from "lodash-es";
 import { flattenDeep } from "lodash-es";
-import getSentences from "../../helpers/sentence/getSentences";
-import { markWordsInSentences } from "../../helpers/word/markWordsInSentences";
+import getSentences from "../helpers/sentence/getSentences";
+import { markWordsInSentences } from "../helpers/word/markWordsInSentences";
 
 /**
  * Calculates the keyword count, takes morphology into account.
